chore: lint and split tests
